hotelReservation/sandboxing/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
PyYAML>=6.0
requests>=2.31.0
